pyproject.toml
[project]
name = "evidence-test"
version = "0.1.0"
description = ""
authors = [
    {name = "Veatec22", email = "wazagiel@outlook.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "duckdb (>=1.3.1,<2.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "requests (>=2.31.0,<3.0.0)",
    "pandas (>=2.0.0,<3.0.0)",
    "gspread (>=6.0.0,<7.0.0)",
    "gspread-dataframe (>=4.0.0,<5.0.0)",
    "oauth2client (>=4.1.3,<5.0.0)",
    "python-dotenv (>=1.0.0,<2.0.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

